requirements.txt
uvicorn[standard]
sqlalchemy
fastapi
google-adk
